==> notebooks/1.0-data-exploration.ipynb <==
import pandas as pd

# Reading File

FILE_PATH = '../data/raw/survey_results_public.csv'
SAVE_PATH = '../data/interim/'

survey = pd.read_csv(FILE_PATH)
survey

# Choose columns for analysis and for recommendations

survey.info()

chosen_analysis_columns = ['MainBranch', 'Employment', 'RemoteWork',
                         'EdLevel', 'YearsCode', 'YearsCodePro', 'DevType',
                         'OrgSize', 'Country', 'CompTotal', 'CompFreq',
                         'LanguageHaveWorkedWith', 'LanguageWantToWorkWith',
                         'DatabaseHaveWorkedWith', 'DatabaseWantToWorkWith',
                         'PlatformHaveWorkedWith', 'PlatformWantToWorkWith',
                         'WebframeHaveWorkedWith', 'WebframeWantToWorkWith',
                         'MiscTechHaveWorkedWith', 'MiscTechWantToWorkWith',
                         'ToolsTechHaveWorkedWith', 'ToolsTechWantToWorkWith',
                         'NEWCollabToolsHaveWorkedWith', 'NEWCollabToolsWantToWorkWith',
                         'OpSysProfessional use', 'VersionControlSystem', 'Age', 'Gender']
survey = survey[chosen_analysis_columns]

survey.info()

chosen_rec_columns = ['MainBranch', 'DevType',
                         'LanguageHaveWorkedWith', 'LanguageWantToWorkWith',
                         'DatabaseHaveWorkedWith', 'DatabaseWantToWorkWith',
                         'PlatformHaveWorkedWith', 'PlatformWantToWorkWith',
                         'WebframeHaveWorkedWith', 'WebframeWantToWorkWith',
                         'MiscTechHaveWorkedWith', 'MiscTechWantToWorkWith',
                         'ToolsTechHaveWorkedWith', 'ToolsTechWantToWorkWith',
                         'NEWCollabToolsHaveWorkedWith', 'NEWCollabToolsWantToWorkWith']

# Explorate chosen analysis columns

for column in survey.columns:
    if survey[column].dtype == 'object':
        print(column, survey[column].unique(),survey[column].dtype)
        print('*' * 100)

# Save chosen columns

chosen_columns = {'analysis': chosen_analysis_columns,
                  'recommendations':chosen_rec_columns}

import pickle

with open(SAVE_PATH + 'chosen_columns.pkl', 'wb') as f:
    pickle.dump(chosen_columns, f)
