==> notebooks/2.0-data-wrangling.ipynb <==
import pickle
import pandas as pd

FILE_PATH = '../data/raw/survey_results_public.csv'
LOAD_PATH = '../data/interim/'

# Load Data

survey = pd.read_csv(FILE_PATH)
with open(LOAD_PATH + 'chosen_columns.pkl', 'rb') as f:
    chosen_columns = pickle.load(f)

survey = survey[chosen_columns['analysis']]

survey.info()

# Wrangle data
most of the columns are multi choose answers, and answerers are splittable with ';', so will list them in list instead of splittable string

def wrangle_column(column_name):
    target_column = survey[column_name]

    # first check if the columns is splittable with ;
    splittable = target_column.str.contains(';').any()

    if not splittable:
        return target_column
    else:
        # replace ', ' with '_' to make list more readable
        target_column = target_column.str.replace(', ', '_')

        # split column
        target_column = target_column.str.split(';')

        # make nullable values empty list
        mask_null = target_column.isnull()
        target_column.loc[mask_null] =target_column.loc[mask_null].apply(lambda x: [])
        return target_column



for column in survey.columns:
    if survey[column].dtype=='object':
        survey[column] = wrangle_column(column)


survey

survey.to_pickle(LOAD_PATH + '2.0-wrangled-data.pkl')
