==> notebooks/3.0-data-cleaning.ipynb <==
import pandas as pd

FILE_PATH = '../data/raw/survey_results_public.csv'
LOAD_PATH = '../data/interim/'
LOAD_DF_NAME = '2.0-wrangled-data.pkl'
SAVE_DF_NAME = '3.0-cleaned-data.pkl'

# Load Data

survey = pd.read_pickle(LOAD_PATH + LOAD_DF_NAME)

survey.info()

# Clean data
From info dataframe, our data is full of nulls and redundant data that we need to handle

splited_columns = (survey.applymap(type) == list).any()
splited_columns = splited_columns[splited_columns].index

def indexing_nulls(column_name, splitted = True):
    if splitted:
        return survey[survey[column_name].apply(lambda x : not x)].index
    else:
        return survey[survey[column_name].isna()].index


## 1. Remove [DevType, LanguageHaveWorkedWith, YearsCode] nulls
First we need to delete any row that doesn't have any DevType or LanguageHaveWorkedWith or YearsCode, because that it's our target column and also we can't make sure of the info of his survey

imp_columns = ['DevType', 'LanguageHaveWorkedWith', 'YearsCode']
for column in imp_columns:
    non_index = indexing_nulls(column, splitted= column in splited_columns)
    survey.drop(non_index, inplace = True)

### Counting nulls after this step

for column in survey.columns:
    if column in splited_columns:
        print(column, ':', len(indexing_nulls(column)))
    else:
        print(column, survey[column].isna().sum())

# 2. replace nuns values in YearsCodePro to be zero

survey['YearsCodePro'] = survey['YearsCodePro'].fillna('0')

# 3. Convert [YearsCode, YearsCodePro] to be int64

replace_dict = {'Less than 1 year' : 1, 'More than 50 years' : 51 }
replace_column = ['YearsCodePro', 'YearsCode']

for column in replace_column:
    survey[column] = survey[column].replace(replace_dict).astype('int64')

# 4. Remove raws that have YearCode equal to zero

zero_code_index = survey[survey['YearsCode']==0].index
survey.drop(zero_code_index, inplace = True)

survey

survey.to_pickle(LOAD_PATH + SAVE_DF_NAME)
