==> notebooks/4.0-data-preprocessing.ipynb <==
import pandas as pd
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.preprocessing import OrdinalEncoder
from sklearn.preprocessing import OneHotEncoder
import numpy as np
import pickle
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import MinMaxScaler

FILE_PATH = '../data/raw/survey_results_public.csv'
LOAD_PATH = '../data/interim/'
LOAD_DF_NAME = '3.0-cleaned-data.pkl'
SAVE_DF_NAME_ANALYSATION = '4.0-preprocessed-data-analysation.pkl'
SAVE_DF_NAME_RECOMMENDATION = '4.0-preprocessed-data_rec.pkl'


# Load Data

survey = pd.read_pickle(LOAD_PATH + LOAD_DF_NAME)

survey

# 1. Encode MainBranch boolean encoding


list(survey.MainBranch.unique())

replace_dict = {'I am not primarily a developer, but I write code sometimes as part of my work' : 0,
                'I am a developer by profession' : 1,
                'I code primarily as a hobby': 0,
                'I used to be a developer by profession, but no longer am': 1,
                'I am learning to code':0}

survey['MainBranch'] = survey['MainBranch'].replace(replace_dict).astype('int32')

### until here preprocessing dataframe for both tasks finished, I will complete processing for recommendation task, and we will need another notebook to complete Analysation Task

survey.to_pickle(LOAD_PATH + SAVE_DF_NAME_ANALYSATION)

### load recommendations columns

with open(LOAD_PATH + 'chosen_columns.pkl', 'rb') as f:
    chosen_columns = pickle.load(f)

survey = survey[chosen_columns['recommendations']]

survey

# 2. Encode [EdLevel, OrgSize] by using Ordinal encoding

def encode_ordinal(column_name, labels):
    o_encoding = OrdinalEncoder(categories= [labels])
    encoded_column = o_encoding.fit_transform(survey[column_name].fillna('0').values[:,np.newaxis]).astype('int32')
    return encoded_column

### 2.1 EdLevel

list(survey['EdLevel'].unique())

Ed_labels = ['0' ,'Something else',  'Primary/elementary school',
         'Secondary school (e.g. American high school, German Realschule or Gymnasium, etc.)',
          'Some college/university study without earning a degree', 'Associate degree (A.A., A.S., etc.)',
           'Bachelor’s degree (B.A., B.S., B.Eng., etc.)',
          'Master’s degree (M.A., M.S., M.Eng., MBA, etc.)',
           'Professional degree (JD, MD, etc.)',
           'Other doctoral degree (Ph.D., Ed.D., etc.)']
survey['EdLevel'] = encode_ordinal('EdLevel', Ed_labels)



### 2.2 OrgSize

org_labels = ['0' ,'I don’t know', 'Just me - I am a freelancer, sole proprietor, etc.','2 to 9 employees',
         '10 to 19 employees', '20 to 99 employees','100 to 499 employees', '500 to 999 employees', '1,000 to 4,999 employees',
         '5,000 to 9,999 employees', '10,000 or more employees']

survey['OrgSize'] = encode_ordinal('OrgSize', org_labels)

# 3. Scale ['YearsCodePro', 'EdLevel', 'OrgSize'] Columns

# Specify the columns you want to transform
columns_to_transform = ['YearsCodePro', 'EdLevel', 'OrgSize']

# Define the transformers for each column
transformers = [
    ('minmax', MinMaxScaler(), columns_to_transform),
]

# Create the ColumnTransformer
ct = ColumnTransformer(transformers)

# Apply the transformations to the specific columns
transformed_columns = ct.fit_transform(survey)
survey[columns_to_transform] = transformed_columns

survey

# 4. Encode[DevType, all the skills columns] by using MultiLabelBinarizer

encoded_dfs = {}
for column in survey.columns:
    if survey[column].dtype=='object':
        binarizer = MultiLabelBinarizer()
        mask = survey[column].notnull()

        #filter by boolean indexing
        arr = binarizer.fit_transform(survey.loc[mask, column])

    #create DataFrame and add missing (NaN)s index values
        encoded_dfs[column] = (pd.DataFrame(arr, index=survey.index[mask], columns=binarizer.classes_)
               .reindex(survey[column].index, fill_value=0))

    else:
        encoded_dfs[column] = survey[column]


survey = pd.concat(encoded_dfs, axis=1)
survey

survey['DevType']

survey.to_pickle(LOAD_PATH + SAVE_DF_NAME_RECOMMENDATION)
