==> notebooks/4.1-data-preprocessing.ipynb <==
import pandas as pd

FILE_PATH = '../data/raw/survey_results_public.csv'
LOAD_PATH = '../data/interim/'
LOAD_DF_NAME = '4.0-preprocessed-data_rec.pkl'
SAVE_DF_NAME = '4.1-preprocessed-data_skills_dev.pkl'


# Load Data

survey = pd.read_pickle(LOAD_PATH + LOAD_DF_NAME)

survey

------

We need skills only and DevType, because we need to recommend skills to Dev Type or vice versa. but first we need to rescale the existence of these skills based on the other feature for example: if one works as Developer, we can deal its skills with the one who practice programming as hoppy

# 1. Rescale with MainBranch
we can rescale non developer skills as half of importance of developer skills

survey['MainBranch'] = survey['MainBranch']['MainBranch'].astype('float32')
survey.loc[survey['MainBranch','MainBranch'] == 0.0,'MainBranch'] = .5

survey

## 2. Rescale [EdLevel, YearsCodePro, OrgSize]: Add some smothing values to avoid zero of MinMaxScaling

survey.loc[:,['OrgSize', 'EdLevel', 'YearsCodePro']] += .1

survey

# 3. Rescale 'WantToWorkWith' skills to be a half of 'HaveWorkedWith' skills

for column in survey.columns.get_level_values(0).unique():
    if 'WantToWorkWith' in column:
        survey[column] = survey[column] * .5

survey['LanguageWantToWorkWith']

# 4. Concat 'WantToWorkWith' with 'HaveWorkedWith'

skills_df = {}
for column in survey.columns.get_level_values(0).unique():
    if 'HaveWorkedWith' in column:
        want_column = column.replace('HaveWorkedWith','WantToWorkWith')
        new_column = column.replace('HaveWorkedWith','s')
        skills_df[new_column] = survey[column] + survey[want_column]

skills_df = pd.concat(skills_df, axis=1)
skills_df

skills_df['Databases']

for skill_col in skills_df.columns.get_level_values(0).unique():
    skills_df[skill_col] = skills_df[skill_col].values * survey['MainBranch'].values

# 5. Rescale this 'skills df' with values of ['MainBranch', 'OrgSize', 'EdLevel', 'YearsCodePro']

rescale_columns = ['MainBranch', 'OrgSize', 'EdLevel', 'YearsCodePro']
for column in rescale_columns:
    for skill_col in skills_df.columns.get_level_values(0).unique():
        skills_df[skill_col] = skills_df[skill_col].values * survey[column].values

skills_df

# 6. Concat skills_df with DevType

dev_df = survey['DevType']
dev_df.columns = pd.MultiIndex.from_product([['DevType'],survey['DevType'].columns])


skills_dev_df = pd.concat([skills_df, dev_df], axis=1, join='inner')

skills_dev_df['DevType']

skills_dev_df.to_pickle(LOAD_PATH + SAVE_DF_NAME)
