==> notebooks/7.0-modaling-base_model.ipynb <==
import sys
import os

SCRIPT_DIR = os.path.dirname(os.path.abspath('src'))
sys.path.append(os.path.dirname(SCRIPT_DIR))

# Load packages
import pandas as pd
import numpy as np

import sklearn.metrics as skm
import sklearn.preprocessing as skp
from sklearn.multioutput import MultiOutputClassifier
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from src.utils import MlflowUtils
from sklearn.pipeline import make_pipeline
from modelingUtlis import ModelingUtils
import plotly.express as px
import plotly.graph_objects as go
import plotly.figure_factory as ff


LOAD_PATH = '../data/interim/'
LOAD_SKILLS_DEV = '6.0-Engineered_data-split_roles-cluster_skills.pkl'
SAVE_DF_NAME = '7.0-Chosen_features_and_roles.pkl'
SAVE_DF_NAME_CSV = '7.0-Chosen_features_and_roles.csv'

# Load Data

skills_dev_df = pd.read_pickle(LOAD_PATH + LOAD_SKILLS_DEV)

skills_dev_df

# 1. Choosing features
we need to choose between original features or clusters features, which one gives us the best results in classification to proceed with them in hero model

we will use class weights formula:
# $$ w_j=\frac{n\_samples}{(n\_classes * n\_samples_j)} $$

basic_model = make_pipeline(skp.StandardScaler(), MultiOutputClassifier(LogisticRegression()))


original_df = skills_dev_df.drop('clustered_skills', level = 0, axis = 1)
original_modeling_utils = ModelingUtils(original_df, basic_model)
original_features_model, classification_report_original_features = original_modeling_utils.train_evaluate_model_features()

drop_columns = original_df.drop('DevType', level=0, axis=1).columns
clustered_df = skills_dev_df.drop(drop_columns, axis = 1)
clustered_modeling_utils = ModelingUtils(clustered_df, basic_model)
clustered_features_model, classification_report_clustered_features = clustered_modeling_utils.train_evaluate_model_features()

classification_report_original_features.loc['Mean',:] = classification_report_original_features.mean()
classification_report_original_features

classification_report_clustered_features.loc['Mean',:] = classification_report_clustered_features.mean()
classification_report_clustered_features


## 1. From the above results we will choose original features as our features
## 2. we need to also Exclude roles that have f1-score less than 0.1

drop_roles = classification_report_original_features[classification_report_original_features['test']['recall'] < .1].index
drop_roles

classification_report_original_features = classification_report_original_features.drop(drop_roles)
classification_report_original_features.loc['Mean',:] = classification_report_original_features.mean()
classification_report_original_features

classification_report_clustered_features = classification_report_clustered_features.drop(drop_roles)
classification_report_clustered_features.loc['Mean',:] = classification_report_clustered_features.mean()
classification_report_clustered_features

job_names = classification_report_clustered_features[:-1].index
job_names = list(job_names)
job_names

# 6. Concat skills_df with DevType

skills_df = skills_dev_df.drop('clustered_skills', axis = 1, level=0).drop('DevType', axis = 1, level = 0)

dev_df = skills_dev_df['DevType'].drop(drop_roles, axis = 1)
dev_df.columns = pd.MultiIndex.from_product([['DevType'],job_names])


skills_dev_df = pd.concat([skills_df, dev_df], axis=1, join='inner')
skills_dev_df

skills_dev_df.to_pickle(LOAD_PATH + SAVE_DF_NAME)


skills_dev_df.to_csv(LOAD_PATH + SAVE_DF_NAME_CSV)

# Save logs

## 1. original features

classification_report_original_features

original_features = original_df.drop('DevType',level=0, axis=1).droplevel(level=0,axis=1).columns
original_modeling_utils.save_results(all_classification_report=classification_report_original_features,
                                     model_name='Logistic-Regression_Original-Features_1.0.0',
                                     data_path=LOAD_SKILLS_DEV, features=original_features)


clustered_features = clustered_df.drop('DevType',level=0, axis=1).droplevel(level=0,axis=1).columns
clustered_modeling_utils.save_results(all_classification_report=classification_report_clustered_features,
                                     model_name='Logistic-Regression_Clustered-Features_1.0.0',
                                     data_path=LOAD_SKILLS_DEV, features=clustered_features)

MlflowUtils.get_runs()

### as we see from the above dataframe, original features outperform clustered features in all metricises especially recall & f1_score that we care more
#### 1. so will use the original features not all the clustered features
#### 2. And we will use These metricises as reference for any further complex model
