==> notebooks/7.1-modeling-XGBoost.ipynb <==
import sys
import os

SCRIPT_DIR = os.path.dirname(os.path.abspath('src'))
sys.path.append(os.path.dirname(SCRIPT_DIR))


import numpy as np
import pandas as pd
import xgboost as xgb

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from src.utils import MlflowUtils
import sklearn.metrics as skm
from sklearn.multioutput import MultiOutputClassifier
from sklearn.model_selection import train_test_split
from sklearn.pipeline import make_pipeline
from modelingUtlis import ModelingUtils


LOAD_PATH = '../data/interim/'
LOAD_SKILLS_DEV = '7.0-Chosen_features_and_roles.pkl'


# Load Data

skills_dev_df = pd.read_pickle(LOAD_PATH + LOAD_SKILLS_DEV)

skills_dev_df

new_pipeline = make_pipeline(StandardScaler(), MultiOutputClassifier(xgb.XGBClassifier(min_child_weight = 1.4, gamma = 2.5, alpha = 1, learning_rate = .1, max_depth = 3, n_estimators = 200)))
XGB_modeling_utils = ModelingUtils(skills_dev_df, new_pipeline)
XGB_features_model, XGB_classification_report = XGB_modeling_utils.train_evaluate_model_features()

XGB_classification_report.loc['Mean',:] = XGB_classification_report.mean()
XGB_classification_report

# save logs

XGB_features = skills_dev_df.drop('DevType',axis=1, level=0).droplevel(level=0,axis=1).columns
XGB_modeling_utils.save_results(all_classification_report=XGB_classification_report,
                                     model_name='XGB_1.0.0',
                                     data_path=LOAD_SKILLS_DEV, features=XGB_features)

MlflowUtils.get_runs()
