==> notebooks/8.0-roles-prediction.ipynb <==
import sys
import os

SCRIPT_DIR = os.path.dirname(os.path.abspath('src'))
sys.path.append(os.path.dirname(SCRIPT_DIR))


from src.utils import MlflowUtils
import pandas as pd
import prediction_utils as utils

MlflowUtils.get_runs()

MlflowUtils.get_run_names()

full_stack_skills = ['HTML/CSS', 'Angular']
php_skills = ['PHP', 'Laravel', 'MariaDB', 'Vue.js', 'HTML/CSS', 'Angular' ]
data_skills = ['Python', 'Scala', 'Hadoop', 'Apache Spark']

features_names, target_names, model  = utils.fetch_best_model_data()
df = utils.prepare_df(data_skills, features_names)
predictions = utils.predict_roles(model, df, target_names)
predictions

predictions['full_stack_PHP']
